Struct try_2::Solution

pub struct Solution;

Implementations

impl Solution

pub fn pentagram_game() -> Vec<Vec<i32>>

Question 1 - Pentagram Game

Enter 1..=12 into pentagram, except 7 and 11. the spins and mirrors are trusted as same. The pentagram and each index are as follow.

_____0___________________4(0)
8__1__7___6_____2(1)___4(1)_2(0)______3(2)
__2____5_____________4(2)_____3(1)
____4____________________3(0)
3_______9________4(3)____________1(0)

The number filled into should be:

0 (4(0)) + 1 (4(1)) + 2 (4(2)) == 4 (3(0)) + 5 (3(1)) + 6 (3(2))
3 (4(3)) + 4 (3(0)) + 5 (3(1)) == 7 (2(0)) + 1 (4(1)) + 8 (2(1))
6 (3(2)) + 7 (2(0)) + 1 (4(1)) == 2 (4(2)) + 4 (3(0)) + 9 (1(0))
8 (2(1)) + 2 (4(2)) + 4 (3(0)) == 5 (3(1)) + 7 (2(0)) + 0 (4(0))

Note: This method using DFS is extremely sloooooowly, but we can filter out some branches for hacking. As a result, 3.55s -> 0.07s.

pub fn red_and_black(room: &Vec<Vec<char>>, pos: (usize, usize)) -> usize

Question 2 - Red and Black

‘.’ is a black block, ‘#’ is a red block, ‘@’ is a human. The human can only move between the black blocks. This method use a simple DFS algorithm, but it’s sloooooowly.

pub fn sliding_puzzle<T: Clone + Eq + Hash>(
src: &Vec<Vec<T>>,
dist: &Vec<Vec<T>>,
nul: T
) -> Option<usize>

Question 3 - Sliding Puzzle Problem

This method using a Two-way BFS which prefered to search the shortest path when both the src and dist are known. Firstly, we check the bound of the board, and then search the null slot to be swapped. Then we need to calculate a “magical number” (a.k.a. inverse number) for verifying whether the board is valid. The detail is that we have to map the src -> dist for the math inverse number. If the puzzle is valid (inverse number (cnt) is the power of 2), we create double deque for Two-way BFS, one start by src and the other start by dist. We can simply setup a hash map to record whether a step is repeated and who passby it. For moving the blocks, the valid directions are filtered by the bound of the board. Due to the borrowed rules, I swap the raw ptr to avoid double mutable borrow (which actually safe sometime).

pub fn solve_n_queens(n: usize) -> Vec<Vec<String>>

Question 4 - Eight Gueens Problem (0 ms 2.3 MB)

Write an algorithm to print all ways of arranging n queens on an n x n chess board so that none of them share the same row, column, or diagonal. In this case, “diagonal” means all diagonals, not just the two that bisect the board.

https://leetcode-cn.com/problems/eight-queens-lcci

This problem can be simply solved by DFS. Firstly, we build a “map” with each line to it’s column, then we fork the “map” each different choose (column). The only thing we should pay attention is cut the fork invalid (line 319 ~ 328). When the recursion on the top (map forked is full), we convert the map to result’s element (line 307 ~ 316).

Issues

Add this if FromIterator didn’t preluded by your compiler.

use core::iter::FromIterator;

If you are using a stable Rust and have some issues with the abs_diff. Simply add this following function into the body, then modify the line 322.

use std::{cmp::Ordering, ops::Sub};

fn abs_diff<T: Default + Ord + Sub<Output = T>>(a: T, b: T) -> T {
    match a.cmp(&b) {
        Ordering::Equal => T::default(),
        Ordering::Greater => a - b,
        Ordering::Less => b - a,
    }
}